# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application ToDo & Co ",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://127.0.0.1:8000/",
            "description": "Projet openclassrooms : Améliorez une application existante de ToDo & Co"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "App\\Controller\\UserController::show",
                "responses": {
                    "200": {
                        "description": "liste des utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Users"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Tasks": {
                "description": "list des taches",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string",
                        "nullable": false
                    },
                    "content": {
                        "type": "string",
                        "nullable": false
                    },
                    "isDone": {
                        "type": "boolean",
                        "nullable": false
                    },
                    "createAt": {
                        "type": "dateTime",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "Users": {
                "description": "list des utilisateurs",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string",
                        "nullable": false
                    },
                    "password": {
                        "type": "string",
                        "nullable": false
                    },
                    "username": {
                        "type": "string",
                        "nullable": false
                    },
                    "roles": {
                        "type": "string",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "NotFound": {
                "description": "La resource n'existe pas",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "La resource n'existe pas"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID de la resource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}